Reorganize some more files
