+++ backend/tpbackend/main.py
@@ -1,22 +1,15 @@
-import uvicorn
 import asyncio
 import os
-from .oblivionis import storage as oblivionisStorage, sync as oblivionisSync
+from .oblivionis import storage as oblivionis_storage, sync as oblivionis_sync
 from .discord.bot import bot
 from tpbackend.storage import db, clean_loop
-from fastapi import FastAPI
-from .api_router import api_router
+from tpbackend.api.api import TimeplayedAPI
 
 # from tpbackend.api import router as api_v1_deprecated
 
 
 async def start_api():
-    app = FastAPI(title="timeplayed")
-    app.include_router(api_router)
-
-    config = uvicorn.Config(app, host="0.0.0.0", port=8000, log_level="info")
-    server = uvicorn.Server(config)
-    await server.serve()
+    await TimeplayedAPI().run()
 
 
 async def start_bot():
@@ -25,12 +18,12 @@ async def start_bot():
 
 async def async_main():
     await asyncio.gather(
-        start_api(), start_bot(), oblivionisSync.sync_loop(), clean_loop()
+        start_api(), start_bot(), oblivionis_sync.sync_loop(), clean_loop()
     )
 
 
 def main():
-    oblivionisStorage.connect_db()
+    oblivionis_storage.connect_db()
     db.connect()
     asyncio.run(async_main())
