+++ backend/tpbackend/user/routes.py
@@ -2,7 +2,7 @@ from tpbackend.activity.query import ActivityQuery
 from tpbackend.utils2 import clamp, parseTS, parse_csv
 from tpbackend.user.query import UserStatsQuery, UserQuery
 from tpbackend.user.models import API_UserWithStats, API_User
-from tpbackend.http_responses import bad_request, not_found
+from tpbackend.api.responses import bad_request, not_found
 import logging
 from fastapi import APIRouter, Path
 from tpbackend.common.types import (
